refactor(store): remove projection runner (moved to nexus-framework)
Deletes runner/, builder, error, and persist modules from nexus-store.
Moves runner tests and persistence tests to nexus-framework.
All 17 framework tests pass.
