+++ crates/nexus-store/src/projection/mod.rs
@@ -1,7 +1,6 @@
 mod pending;
 mod persisted;
 mod projector;
-pub mod runner;
 mod store;
 #[cfg(feature = "testing")]
 mod testing;
